make fast cursor type changes
Until now, the cursor changes very slowly when exiting insert mode. It
remains as a thin line for a second or so, before going to the block
shape it is supposed to have. With this, it changes almost immediately.

Adopted from stackoverflow thread:
https://stackoverflow.com/questions/4777950/vim-change-block-cursor-when-in-insert-mode

I skipped the listchars setting, which did not seem related to faster
cursor switches.